requirements.txt
Django~=4.2.3
tox~=4.6.4
pytest-django~=4.5.2
model_bakery~=1.12.0
pytest-cov~=4.0.0
django-phonenumber-field~=7.1.0
phonenumbers~=8.13.16